Refactor package name from org.usfirst[...].robot to org.robockets.robot.
The robockets own robockets.org and so it makes sense to use org.robockets. The
robot makes sense because it's robot code.

It should be considered whether or not we want to append the year at the end of
the package name because in the future the package cannot be named
org.robockets.robot because that's already taken!