Move number of barrels to RobotMap and define PID methods.
I moved the variable of number of barrels to RobotMap as it describes the
physical hardware on the robot. I defined PID methods such as resetPID,
spinGatlin for aligning the PID, and other goodies.

And yes I now know that is not how you spell a Gatling gun.

It seems I will need to refactor.