Add PopCannon command.
This command simply opens and closes the Solenoid for a designated amount
of time in order to fire a tshirt. This command also checks to see if the cannon
is aligned. If it is not it will end and make sure we don't mess up, how
wonderful!

As a note to the future, javadocs aren't put in yet, that is a needed addition.