Remove unused variable and allow PID to be disabled.
numberOfBarrels is no longer needed, all the logic for barrels and distance
will be done in the RobotMap. Another change is that I have not allowed PID
to be disabled by flipping the value of the PID variable (it's publicly visible).
There should probably be a method for it as changing values is too low-level.
Maybe that will be another commit.

It disables it by checking the PID variable in the usePIDOutput callback, if the
value of PID is false, then it won't do anything to the motor.

It should be noted that this may cause the PID when resumed to freak out.
Maybe returnPIDInput should have the same check.