Add safety system for cannon
Basically I copy/pasted stuff I wrote in 2015-Robot-Code, then I set the PopCannon command to not fire if there is anything within 2 meters (200 cm) of the distance sensor. For now, the formula is for a GP2D120XJ00F (ShortDistanceSensor) which will not work in real life for this purpose. We will need to find a better distance sensor. Also, the port for the distance sensor is set to -1 which _may_ crash the program if we try to run it.